Consolidate RT memory setup into rt_memory
Move the mlockall()/mallopt()/RLIMIT_MEMLOCK-logging orchestration out of
control_node.cpp into rt_memory.cpp behind configure_realtime_memory(),
per PR #12 review. control_node's control-loop thread now makes a single
call. The pure, unit-tested helpers stay header-only and ROS-free
(rclcpp::Logger is forward-declared), so test_rt_memory still links
without rclcpp.

Refs SW-933
